requirements.txt
black==22.1.0
boto3==1.21.14
botocore==1.24.14
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
et-xmlfile==1.1.0
greenlet==1.1.2
isort==5.10.1
jmespath==0.10.0
# jwt==1.3.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mypy-extensions==0.4.3
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.1
pathspec==0.9.0
platformdirs==2.5.1
psycopg2==2.9.3
aws-psycopg2==1.3.8
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
PyJWT==2.4.0